students/k33391/Igor_Tulukin/Lr3/task2/parser/requirements.txt
aiohttp==3.10.5
beautifulsoup4==4.12.3
fastapi==0.115.0
psycopg2_binary==2.9.9
python-dotenv==1.0.1
